requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & OAuth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.1.1
authlib==1.2.1
httpx==0.25.2

# Email protocols
aiosmtpd==1.4.4
imaplib2==3.6
poplib==0.1

# Email processing
email-validator==2.1.0
dnspython==2.4.2

# Calendar and address book
icalendar==5.0.7
vobject==0.9.6.1

# Web interface
jinja2==3.1.2
aiofiles==23.2.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3 